Make it possible to set flag values to negative numbers.

+++ src/main/java/net/aufdemrand/denizen/flags/FlagManager.java
@@ -503,7 +503,7 @@ public class FlagManager {
         public double asDouble() {
             adjustIndex();
             try {
-                return Double.valueOf(values.get(index)).intValue();
+                return Double.valueOf(values.get(index));
             } catch (Exception e) { return 0; }
         }
